Refactor to allow custom link implementation, deprecate Anchor
