+++ src/components/Link/Link.md
@@ -1,4 +1,9 @@
-Creates a local or remote link using React Router's Link component. External / internal is inferred by the URL provided to `to`.
+Creates a local or remote link using React Router's Link component. External / internal is inferred by the URL provided to the `to` prop.
+
+Other supported props:
+
+* `appearFocused`: Allows you to programmatically control whether the link is styled as focused. Useful for advanced keyboard navigation scenarios.
+* `Link.Icon` takes the `icon` property, which is the name of the icon to use.
 
 **Basic Links**
 ```js
@@ -11,6 +16,56 @@ Creates a local or remote link using React Router's Link component. External / i
   <Link onClick={() => alert('clicked')}>Link with click handler</Link>
 </div>
 ```
+
+### Customize `Link` with the `LinkImplementation` Component Contract
+
+Link is designed to be flexible and style a user-defined Link component. By default, it ships with the standard HTML `<a>` element. You can provide your own LinkImplementation by doing the following:
+
+#### 1. Implement a component which takes the correct props
+
+A `LinkImplementation` component will be passed all props provided to `<Link>` from the outside, with the exception of a few blacklisted internal props: `appearFocused`, `icon`, `external`, and `passAllPropsToImplementation`.
+
+```js static
+// note: something like this is of course not very useful.
+const CustomLinkImplementation = ({ children, to, ...rest }) => (
+  <div onClick={() => window.location.href = to}>{children}</div>
+);
+```
+
+#### 2. Pass your component as a prop to `Link`
+
+```js static
+const CustomLink = withProps({ LinkImplementation: CustomLinkImplementation })(
+  Link
+);
+```
+
+#### 3. Optional: disable prop blacklist
+
+You can optionally pass the `passAllPropsToImplementation` prop through your `<Link>` to disable this prop blacklisting.
+
+### Use with React Router
+
+To use `Link` with React Router's `Link`, you need to provide a custom implementation. Define a `Link` component within your local project which passes a component that satisfies the `LinkImplementation` contract:
+
+```js static
+import React from 'react';
+import { Link as RouterLink } from 'react-router-dom';
+import { Link } from '@bandwidth/shared-components';
+import { withProps } from 'recompose';
+
+const RouterLinkImplementation = ({ children, ...rest }) => (
+  <RouterLink {...rest}>{children}</RouterLink>
+);
+
+export default withProps({ LinkImplementation: RouterLinkImplementation })(
+  Link
+);
+```
+
+You can use this React-Router-powered Link for both internal and external links. It will automatically apply the appropriate props to force the React Router Link component to route externally if it detects a URL with a protocol (`http(s)://`)
+
+### More Link Styles
 
 **Text Link Styles**
 ```js
