+++ src/components/Link/styles/NativeALinkImplementation.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default ({ children, to, ...rest }) => (
+  <a {...rest} href={to}>
+    {children}
+  </a>
+);
