+++ src/components/Link/styles/TransparentLink.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import DomSafeLink from './PropBlacklistLink';
+import DomSafeLink from './RenderLinkImplementation';
 
 export default styled(DomSafeLink)`
   text-decoration: none;
